requirements.txt
requests==2.32.3
flask==3.1.0
websocket-client==1.8.0
